requirements.txt
Pillow
matplotlib
numpy
plotly
kaleido